fix(planning_test_utils): rename header directory to match project name
